Refactor contact handling to associate contacts with users, enforce permissions on updates and deletions, and improve token validation error handling

+++ controllers/contactController.js
@@ -3,18 +3,19 @@ const Contact = require("../models/contactModel");
 
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contact = await Contact.findById(req.params.id);
-    if(!contact){
-        res.status(404);
-        throw new Error("Contacts not found");
-    }
-    res.status(200).json(contact);
+  const contacts = await Contact.find({ user_id: req.user.id });
+  res.status(200).json(contacts);
 });
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -30,6 +31,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to delete other user contacts");
     }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact);
